calculating Moving averages - package updated
