+++ algorithms-modules/algorithms-miscellaneous-8/pom.xml
@@ -13,4 +13,12 @@
         <version>1.0.0-SNAPSHOT</version>
     </parent>
 
+
+    <dependencies>
+        <dependency>
+            <groupId>org.apache.commons</groupId>
+            <artifactId>commons-math3</artifactId>
+            <version>${commons-math3.version}</version>
+        </dependency>
+    </dependencies>
 </project>
