+++ algorithms-modules/algorithms-miscellaneous-8/src/main/java/com/baeldung/movingaverages/ExponentialMovingAverage.java
@@ -1,4 +1,4 @@
-package com.baeldung;
+package com.baeldung.movingaverages;
 
 public class ExponentialMovingAverage {
