+++ algorithms-modules/algorithms-miscellaneous-8/src/main/java/com/baeldung/movingaverages/MovingAverageWithStreamBasedApproach.java
@@ -1,4 +1,4 @@
-package com.baeldung;
+package com.baeldung.movingaverages;
 
 import java.util.stream.DoubleStream;
