+++ algorithms-modules/algorithms-miscellaneous-8/src/test/java/com/baeldung/movingaverages/ExponentialMovingAverageUnitTest.java
@@ -1,7 +1,8 @@
-package com.baeldung;
+package com.baeldung.movingaverages;
+
+import org.junit.Test;
 
 import static org.junit.Assert.assertEquals;
-import org.junit.Test;
 
 public class ExponentialMovingAverageUnitTest {
