+++ algorithms-modules/algorithms-miscellaneous-8/src/test/java/com/baeldung/movingaverages/MovingAverageWithApacheCommonsMathUnitTest.java
@@ -1,6 +1,7 @@
-package com.baeldung;
+package com.baeldung.movingaverages;
 
 import org.junit.Test;
+
 import static org.junit.Assert.assertEquals;
 
 public class MovingAverageWithApacheCommonsMathUnitTest {
