+++ algorithms-modules/algorithms-miscellaneous-8/src/test/java/com/baeldung/movingaverages/MovingAverageWithStreamBasedApproachUnitTest.java
@@ -1,8 +1,8 @@
-package com.baeldung;
+package com.baeldung.movingaverages;
 
 import org.junit.Test;
 
-import static org.junit.jupiter.api.Assertions.*;
+import static org.junit.jupiter.api.Assertions.assertEquals;
 
 public class MovingAverageWithStreamBasedApproachUnitTest {
