+++ algorithms-modules/pom.xml
@@ -26,7 +26,7 @@
         <module>algorithms-searching</module>
         <module>algorithms-sorting</module>
         <module>algorithms-sorting-2</module>
-        <module>algorithms-miscellaneous-9</module>
+        <module>algorithms-miscellaneous-8</module>
     </modules>
 
     <properties>
